split up Announcements ViewModel
