+++ app/src/main/java/de/xikolo/viewmodels/AnnouncementsViewModel.kt
@@ -4,50 +4,26 @@ import androidx.lifecycle.LiveData
 import de.xikolo.models.Announcement
 import de.xikolo.models.dao.AnnouncementsDao
 import de.xikolo.network.jobs.ListCourseAnnouncementsJob
-import de.xikolo.network.jobs.ListGlobalAnnouncementsJob
-import de.xikolo.network.jobs.UpdateAnnouncementVisitedJob
 import de.xikolo.viewmodels.base.BaseViewModel
 
-class AnnouncementsViewModel(val courseId: String? = null) : BaseViewModel() {
+class AnnouncementsViewModel(val courseId: String) : BaseViewModel() {
 
     private val announcementsDao = AnnouncementsDao(realm)
 
-    val courseAnnouncements: LiveData<List<Announcement>> by lazy {
-        if(courseId != null) {
-            announcementsDao.getAnnouncementsForCourse(courseId)
-        } else {
-            throw IllegalArgumentException("courseId must not be null to access this field")
-        }
-    }
-
-    val globalAnnouncements: LiveData<List<Announcement>> by lazy {
-        announcementsDao.getGlobalAnnouncements()
+    val announcements: LiveData<List<Announcement>> by lazy {
+        announcementsDao.getAnnouncementsForCourse(courseId)
     }
 
     override fun onFirstCreate() {
-        if(courseId != null) {
-            requestCourseAnnouncementList(courseId, false)
-        }
-        requestGlobalAnnouncementList(false)
+        requestCourseAnnouncementList(courseId, false)
     }
 
     override fun onRefresh() {
-        if(courseId != null) {
-            requestCourseAnnouncementList(courseId, true)
-        }
-        requestGlobalAnnouncementList(true)
+        requestCourseAnnouncementList(courseId, true)
     }
 
     private fun requestCourseAnnouncementList(courseId: String, userRequest: Boolean) {
         ListCourseAnnouncementsJob(courseId, userRequest, networkState).run()
-    }
-
-    private fun requestGlobalAnnouncementList(userRequest: Boolean) {
-        ListGlobalAnnouncementsJob(userRequest, networkState).run()
-    }
-
-    fun updateAnnouncementVisited(announcementId: String) {
-        UpdateAnnouncementVisitedJob.schedule(announcementId)
     }
 
 }
